Changed Cytomander's wording
"Venting" usually implies taking one's frustration out on someone/thing verbally, not physically. I think this addition will make it sound better.
I tried to come up with a synonym for venting that worked, but I honestly couldn't think of any.
If you do, feel free to replace this with it.